pyproject.toml
[tool.poetry]
name = "currrency-rate"
version = "0.1.0"
description = ""
authors = ["korchaginks <kkonstantinserg@gmail.com>"]
readme = "README.md"
packages = [{include = "currrency_rate"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
